@@ -1,6 +1,6 @@
 {
-    "name": "@bhplugin/vue3-datatable",
-    "version": "2.0.2",
+    "name": "@anemo27/vue3-datatable",
+    "version": "1.0.0",
     "description": "Vue3 Datatable - fully customizable & easy to use datatable library",
     "private": false,
     "type": "module",
@@ -18,18 +18,18 @@
         "./dist/style.css": "./dist/style.css"
     },
     "author": {
-        "name": "Bhavesh Patel",
-        "email": "bhavesh.patel200@gmail.com"
+        "name": "Anemo27",
+        "email": "premieofficial@gmail.com"
     },
     "repository": {
         "type": "git",
-        "url": "git+https://github.com/bhaveshpatel200/vue3-datatable.git"
+        "url": "git+https://github.com/Anemo27/vue3-datatable.git"
     },
     "bugs": {
-        "url": "https://github.com/bhaveshpatel200/vue3-datatable/issues"
+        "url": "https://github.com/Anemo27/vue3-datatable/issues"
     },
-    "homepage": "https://github.com/bhaveshpatel200/vue3-datatable#readme",
-    "funding": "https://github.com/sponsors/bhaveshpatel200",
+    "homepage": "https://github.com/Anemo27/vue3-datatable#readme",
+    "funding": "https://github.com/sponsors/Anemo27",
     "keywords": [
         "vue3",
         "vue",
@@ -58,19 +58,23 @@
     },
     "scripts": {
         "dev": "vite",
-        "build": "vite build && vue-tsc --emitDeclarationOnly && npm run tailwind:build",
+        "build": "vite build && vue-tsc --emitDeclarationOnly && pnpm run tailwind:build && pnpm run copy-css",
         "preview": "vite preview",
-        "tailwind": "tailwindcss -i ./src/assets/css/tailwind.css -o ./dist/style.css --watch",
-        "tailwind:build": "tailwindcss -i ./src/assets/css/tailwind.css -o ./dist/style.css --minify"
+        "tailwind": "npx @tailwindcss/cli -i ./src/assets/css/tailwind.css -o ./dist/style.css --watch",
+        "tailwind:build": "npx @tailwindcss/cli -i ./src/assets/css/tailwind.css -o ./dist/style.css --minify",
+        "copy-css": "copy dist\\style.css dist\\vue3-datatable.css"
     },
     "devDependencies": {
-        "@types/node": "^18.11.9",
-        "@vitejs/plugin-vue": "^3.0.0",
-        "autoprefixer": "^10.4.8",
-        "typescript": "^4.6.4",
-        "vite": "^3.0.0",
-        "vue-tsc": "^0.38.4",
-        "vue": "^3.5.10"
+        "@tailwindcss/postcss": "^4.1.13",
+        "@types/node": "^24.5.0",
+        "@vitejs/plugin-vue": "^6.0.1",
+        "autoprefixer": "^10.4.21",
+        "postcss": "^8.5.6",
+        "tailwindcss": "^4.1.13",
+        "typescript": "^5.9.2",
+        "vite": "^7.1.5",
+        "vue": "^3.5.21",
+        "vue-tsc": "^3.0.7"
     },
     "peerDependencies": {
         "vue": ">=3.0.0"
