@@ -1,9 +1,5 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
-    corePlugins: {
-        preflight: false,
-    },
-    prefix: 'bh-',
+export default {
     content: ['./index.html', './src/**/*.{vue,js,ts,jsx,tsx}'],
     theme: {
         fontFamily: {
